# protocol/openapi.yaml
openapi: 3.0.3
info:
  title: Remote File Explorer — Host Agent API
  version: 1.17.0
  description: >
    Contract between the Flutter mobile app and the Go host agent. JSON for control and
    metadata; binary streams for file content. All routes except /pair, /register,
    /login, /auth/challenge, and /health require a bearer device token. Clients may
    send an optional `X-RFE-Client-Version` header (e.g.
    "1.10.0+18") on every authenticated request; the agent records this plus the
    caller's network address against the device row. Admin devices see those fields
    for all devices via GET /devices (see Device.lastAddress / Device.lastVersion);
    ordinary paired devices receive only their own minimal device record.

    A device's token is "admin" (Device.viaLogin=true) if it was obtained via /login or
    /register (proof of the single account's password) rather than /pair (a one-time
    code meant for an ordinary device like the phone app). Admin devices may manage
    OTHER devices: POST /pairing/generate, and PATCH/DELETE /devices/{id} targeting a
    device other than themselves. An ordinary (non-admin) device may still manage only
    ITSELF via DELETE /devices/{id}.

    Pair, register, and login all require device-identity proof-of-possession: fetch a
    nonce from POST /auth/challenge, sign it with the device's permanent Ed25519 key,
    and send the (devicePublicKey, nonce, signature) triple. The agent pins the public
...
servers:
...
    variables:
...
security:
  - deviceToken: []
paths:
...
  /share/{token}:
    get:
      summary: Fetch a shared file (UNAUTHENTICATED, single-use)
      description: >
        The only unauthenticated route in the entire agent besides /health
        and /pair. token is single-use: the first successful (or attempted)
        fetch deletes it. Requests are limited to 10/min per source IP; the
        agent keeps at most 4096 IP buckets and rejects new IPs while the
        table is full. The opened file must still be regular and at most
        500 MiB. Returns 404 for both "expired" and "never existed" — the
        response never reveals which.
      security: []
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        "200": { description: File bytes, content: { application/octet-stream: { schema: { type: string, format: binary } } } }
        "404": { $ref: "#/components/responses/Error" }
        "413":
          description: The file exceeds the 500 MiB share limit. The single-use token is consumed. code=FILE_TOO_LARGE.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "429":
          description: Too many share requests from this source IP, or the bounded rate-limit table is full. code=RATE_LIMITED.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    delete:
      summary: Revoke an active share link
      description: >
        For DELETE the {token} path segment is the SHA-256 HASH of the token
        (as returned by POST /share/mint's tokenHash field), not the raw
        token used by GET.

        Scoped to the link's owner: only the device that minted it, or an
        admin device, may revoke it. Anyone else gets 404 — the same response
        an unknown hash gets, so a foreign hash is never confirmed. Links
        minted before the agent recorded owners are admin-only.
      security: [{ deviceToken: [] }]
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Revoked }
        "401": { $ref: "#/components/responses/Error" }
        "404":
          description: >
            No such link, or it belongs to another device. code=NOT_FOUND.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /share:
    get:
      summary: List active (unexpired) share links
      description: >
        Scoped to the caller: an admin device (authenticated via /login or
        /register) sees every active link; a code-paired device sees only the
        links it minted. Links minted before the agent recorded owners appear
        for admins only.
      security: [{ deviceToken: [] }]
      responses:
        "200":
          description: Active share links
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ShareLinkSummary" }
        "401": { $ref: "#/components/responses/Error" }
        "403":
          description: >
            Caller has no device identity at all. code=FORBIDDEN.
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    ShareLinkSummary:
      type: object
      properties:
        tokenHash: { type: string }
        path: { type: string }
        expiresAt: { type: integer, format: int64, description: "Unix seconds." }
...
    Error:
      type: object
      description: >
        Uniform error envelope returned on non-2xx responses. The `code` field is
        the authoritative machine-readable vocabulary the agent emits (see enum);
        `message` is a human-readable detail string and may change without notice.
      properties:
        code:
          type: string
          example: PATH_NOT_FOUND
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - READ_ONLY
            - FORBIDDEN
            - PATH_NOT_FOUND
            - CONFLICT
            - CHUNK_HASH_MISMATCH
            - HASH_MISMATCH
            - NOT_AVAILABLE
            - THUMB_BUSY
            - INVALID_CODE
            - DELETE_FAILED
            - RATE_LIMITED
            - RESOURCE_LIMIT
            - TRANSFER_ACTIVE
            - TRANSFER_NOT_OPEN
            - PAYLOAD_TOO_LARGE
            - NOT_FOUND
            - INTERNAL
          description: >
            - `BAD_REQUEST` (400): malformed or missing required fields.
            - `UNAUTHORIZED` (401): missing/invalid/revoked bearer token.
            - `READ_ONLY` (403): write rejected — the agent is in global
              read-only mode OR the calling device has the per-device read-only
              flag set (#8).
            - `FORBIDDEN` (403): path escapes the configured root jail; or a
              /devices/{id} DELETE/PATCH or /pairing/generate call that
              requires an admin device (Device.viaLogin=true) came from a
              non-admin device.
            - `PATH_NOT_FOUND` (404): path does not exist.
            - `NOT_FOUND` (404): PATCH /devices/{id} — no device with that id.
            - `CONFLICT` (409): destination already exists.
            - `CHUNK_HASH_MISMATCH` (409): per-chunk SHA-256 did not match
              X-Chunk-Sha256.
            - `HASH_MISMATCH` (422): whole-file SHA-256 did not match on complete.
            - `NOT_AVAILABLE` (404): no thumbnail/preview available for this file.
            - `THUMB_BUSY` (429): thumbnail request capacity is full; retry after the `Retry-After` interval.
            - `INVALID_CODE` (401): pairing code invalid or expired.
            - `DELETE_FAILED`: per-item batch-delete failure (returned inside a
              BatchResult entry, not as a top-level HTTP error).
            - `RATE_LIMITED` (429): too many /pair attempts (10/min).
            - `RESOURCE_LIMIT` (429): upload admission reached the per-device
              (4 sessions / 64 GiB) or host-wide (16 sessions / 128 GiB)
              session or reserved-byte quota. Retry after completing or
              deleting an existing session; no fixed Retry-After is provided.
            - `TRANSFER_ACTIVE` (409): an operation conflicts with active
              session work; completion is exclusive, and DELETE is rejected
              while another request holds the session.
            - `TRANSFER_NOT_OPEN` (409): the session is completed or failed
              and no longer accepts chunks or completion.
            - `PAYLOAD_TOO_LARGE` (413): chunk body exceeds the session's chunkSize.
            - `INTERNAL` (500): unexpected server error.
        message: { type: string }
        details: { type: object, additionalProperties: true }